+++ CHANGELOG.md
@@ -5,7 +5,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 0.1.2
 ### Added
-*
+* Organization assets
+* Asset distribution
 
 ## 0.1.1 - 2018-05-22
 ### Added